+++ tests/rules/no-duplicate-definitions.test.js
@@ -153,10 +153,8 @@ ruleTester.run("no-duplicate-definitions", rule, {
 				},
 			],
 		},
-		// This test case is skipped for non-Node environments like Bun
-		...(typeof process !== "undefined" &&
-		process.release?.name === "node" &&
-		!process.versions?.bun
+		// This test case is skipped when running on Bun
+		...(!process.versions.bun
 			? [
 					{
 						code: `
@@ -185,7 +183,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 			errors: [
 				{
 					messageId: "duplicateDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 3,
 					column: 1,
 					endLine: 3,
@@ -204,7 +207,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 			errors: [
 				{
 					messageId: "duplicateDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 3,
 					column: 1,
 					endLine: 3,
@@ -212,7 +220,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 				},
 				{
 					messageId: "duplicateDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 4,
 					column: 1,
 					endLine: 4,
@@ -220,7 +233,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 				},
 				{
 					messageId: "duplicateDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 5,
 					column: 1,
 					endLine: 5,
@@ -237,7 +255,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 			errors: [
 				{
 					messageId: "duplicateDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "Mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 3,
 					column: 1,
 					endLine: 3,
@@ -254,7 +277,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 			errors: [
 				{
 					messageId: "duplicateDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 3,
 					column: 1,
 					endLine: 3,
@@ -278,7 +306,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 			errors: [
 				{
 					messageId: "duplicateDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 3,
 					column: 1,
 					endLine: 3,
@@ -295,7 +328,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 			errors: [
 				{
 					messageId: "duplicateFootnoteDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 3,
 					column: 1,
 					endLine: 3,
@@ -314,7 +352,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 			errors: [
 				{
 					messageId: "duplicateFootnoteDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 3,
 					column: 1,
 					endLine: 3,
@@ -322,7 +365,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 				},
 				{
 					messageId: "duplicateFootnoteDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 4,
 					column: 1,
 					endLine: 4,
@@ -330,7 +378,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 				},
 				{
 					messageId: "duplicateFootnoteDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 5,
 					column: 1,
 					endLine: 5,
@@ -347,7 +400,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 			errors: [
 				{
 					messageId: "duplicateFootnoteDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "Mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 3,
 					column: 1,
 					endLine: 3,
@@ -371,7 +429,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 			errors: [
 				{
 					messageId: "duplicateFootnoteDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 3,
 					column: 1,
 					endLine: 3,
@@ -397,7 +460,12 @@ ruleTester.run("no-duplicate-definitions", rule, {
 			errors: [
 				{
 					messageId: "duplicateDefinition",
-					data: { identifier: "mercury" },
+					data: {
+						identifier: "mercury",
+						label: "mercury",
+						firstLine: "2",
+						firstLabel: "mercury",
+					},
 					line: 12,
 					column: 1,
 					endLine: 12,